Mul-AE_v2/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=1.4.0=pypi_0
albumentations=1.3.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
blas=1.0=mkl
brotli=1.0.9=h166bdaf_8
brotli-bin=1.0.9=h166bdaf_8
bzip2=1.0.8=h7f98852_4
ca-certificates=2023.5.7=hbcca054_0
cachetools=5.3.1=pypi_0
certifi=2023.5.7=pyhd8ed1ab_0
charset-normalizer=3.1.0=pyhd8ed1ab_0
contourpy=1.1.0=pypi_0
cudatoolkit=11.3.1=h9edb442_11
cycler=0.11.0=pypi_0
einops=0.4.1=pypi_0
ffmpeg=4.3=hf484d3e_0
fonttools=4.40.0=pypi_0
freetype=2.12.1=hca18f0e_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=2.20.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.54.2=pypi_0
hydra-core=1.0.3=pypi_0
idna=3.4=pyhd8ed1ab_0
imageio=2.31.1=pypi_0
importlib-metadata=6.6.0=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.2.0=pypi_0
jpeg=9e=h0b41bf4_3
kiwisolver=1.4.4=pypi_0
lame=3.100=h166bdaf_1003
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libbrotlicommon=1.0.9=h166bdaf_8
libbrotlidec=1.0.9=h166bdaf_8
libbrotlienc=1.0.9=h166bdaf_8
libdeflate=1.17=h0b41bf4_0
libffi=3.4.2=h7f98852_5
libgcc-ng=13.1.0=he5830b7_0
libgomp=13.1.0=he5830b7_0
libiconv=1.17=h166bdaf_0
libnsl=2.0.0=h7f98852_0
libpng=1.6.39=h753d276_0
libsqlite=3.42.0=h2797004_0
libstdcxx-ng=13.1.0=hfd8a6a1_0
libtiff=4.5.0=h6adf6a1_2
libuuid=2.38.1=h0b41bf4_0
libuv=1.44.2=h166bdaf_0
libwebp-base=1.3.0=h0b41bf4_0
libxcb=1.13=h7f98852_1004
libzlib=1.2.13=hd590300_5
markdown=3.4.3=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.6.0=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h95df7f1_0
mkl_fft=1.3.1=py38h8666266_1
mkl_random=1.2.2=py38h1abd341_0
ncurses=6.4=hcb278e6_0
nettle=3.6=he412f7d_0
networkx=3.1=pypi_0
numpy=1.23.1=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.1.2=pypi_0
opencv-python-headless=4.7.0.72=pypi_0
openh264=2.1.1=h780b84a_0
openjpeg=2.5.0=hfec8fc6_2
openssl=3.1.1=hd590300_1
packaging=23.1=pypi_0
pandas=1.5.0=pypi_0
pillow=9.4.0=py38hde6dc18_1
pip=23.1.2=pyhd8ed1ab_0
protobuf=3.19.6=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.8.16=he550d4f_1_cpython
python-dateutil=2.8.2=pypi_0
python_abi=3.8=3_cp38
pytorch=1.11.0=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2023.3=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
qudida=0.0.4=pypi_0
readline=8.2=h8228510_1
requests=2.31.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
setuptools=67.7.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
tensorboard=2.10.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2023.4.12=pypi_0
timm=0.6.7=pypi_0
tk=8.6.12=h27826a3_0
torchaudio=0.11.0=py38_cu113
torchvision=0.12.0=py38_cu113
typing_extensions=4.6.3=pyha770c72_0
urllib3=1.26.16=pypi_0
werkzeug=2.3.6=pypi_0
wheel=0.40.0=pyhd8ed1ab_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.6=h166bdaf_0
zipp=3.15.0=pypi_0
zlib=1.2.13=hd590300_5
zstd=1.5.2=h3eb15da_6
